Fileupload auf task create

+++ src/SuperSmart/SuperSmart.Core/Data/ViewModels/CreateDocumentViewModel.cs
@@ -13,7 +13,7 @@ namespace SuperSmart.Core.Data.ViewModels
         [Required(ErrorMessage = "Please enter a filename")]
         public string FileName { get; set; }
         [Required(ErrorMessage = "Please enter a taskid")]
-        public Int32 TaskId { get; set; }
+        public long TaskId { get; set; }
     }
 }
